requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
awsebcli==3.20.3
backports.zoneinfo==0.2.1
botocore==1.23.51
cachetools==4.2.4
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0
django-heroku==0.3.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
flatbuffers==2.0
frozenlist==1.2.0
future==0.16.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.9.0
jaraco.context==4.1.1
jmespath==0.10.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
more-itertools==8.12.0
multidict==5.2.0
numpy==1.21.4
oauthlib==3.1.1
openai==0.11.5
openpyxl==3.0.9
opt-einsum==3.3.0
pandas==1.3.5
pandas-stubs==1.2.0.39
pathspec==0.9.0
paystackapi==2.0.0
protobuf==3.19.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
semantic-version==2.8.5
six==1.14.0
sklearn==0.0
sqlparse==0.4.2
static3==0.7.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.3
typing-extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
wcwidth==0.1.9
Werkzeug==2.0.2
whitenoise==6.0.0
wolframalpha==5.0.0
wolframclient==1.1.7
wrapt==1.13.3
xmltodict==0.12.0
yarl==1.7.2
zipp==3.6.0
